+++ Common/src/main/java/de/markusbordihn/easynpc/entity/easynpc/ai/goal/RandomStrollAroundHomeGoal.java
@@ -44,7 +44,7 @@ public class RandomStrollAroundHomeGoal<T extends EasyNPC<?>> extends RandomStro
   public boolean canContinueToUse() {
     return !this.mob.getNavigation().isDone()
         && !this.mob.isVehicle()
-        && this.mob.getTarget() == null;
+        && (!this.mob.isAggressive() || this.mob.getTarget() == null);
   }
 
   @Override